Move build_s3_name_reference_edges outside local_graph
